chore: improve the structure of project
