+++ .idea/workspace.xml
@@ -1,8 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
   <component name="ChangeListManager">
-    <list default="true" id="d5b57014-2c10-4ada-a94c-343e7085c1d7" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/src/components/mainview/MyMainView.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/mainview/MyMainView.jsx" afterDir="false" />
+    <list default="true" id="d5b57014-2c10-4ada-a94c-343e7085c1d7" name="Changes" comment="Style: rebuild whole style.">
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/App.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/assets/img/[Originals]/noDataImage.png" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/footer/MyFooter.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/footer/MyFooter.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/main/MyMain.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/main/MyMain.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/mainview/MyMainView.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/content/MyContent.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/mainview/MyMainView.module.css" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/content/MyContent.module.css" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/menu/MyMenu.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/sidebar/MySidebar.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/menu/MyMenu.module.css" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/sidebar/MySidebar.module.css" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -10,6 +21,11 @@
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="main" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
   <component name="HighlightingSettingsPerFile">
@@ -19,13 +35,54 @@
     <setting file="file://$PROJECT_DIR$/src/API/ReportService.js" root0="FORCE_HIGHLIGHTING" />
     <setting file="file://$PROJECT_DIR$/src/API/Service.json" root0="FORCE_HIGHLIGHTING" />
     <setting file="file://$PROJECT_DIR$/src/components/UI/icon/Icon.jsx" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/components/content/MyContent.jsx" root0="FORCE_HIGHLIGHTING" />
     <setting file="file://$PROJECT_DIR$/src/components/main/MyMain.jsx" root0="FORCE_HIGHLIGHTING" />
-    <setting file="file://$PROJECT_DIR$/src/components/mainview/MyMainView.jsx" root0="FORCE_HIGHLIGHTING" />
-    <setting file="file://$PROJECT_DIR$/src/components/menu/MyMenu.jsx" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/components/sidebar/MySidebar.jsx" root0="FORCE_HIGHLIGHTING" />
   </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2H4nbF2WPBCBKzwS34XtZQJmMSz" />
   <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "WebServerToolWindowFactoryState": "false",
+    "last_opened_file_path": "C:/Users/AlexInter/Workspace/ZohoReport/frontend",
+    "node.js.detected.package.eslint": "true",
+    "node.js.detected.package.tslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "node.js.selected.package.tslint": "(autodetect)",
+    "nodejs_package_manager_path": "npm",
+    "ts.external.directory.path": "C:\\Program Files\\JetBrains\\WebStorm 2022.1.3\\plugins\\javascript-impl\\jsLanguageServicesImpl\\external",
+    "vue.rearranger.settings.migration": "true"
+  }
+}]]></component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="d5b57014-2c10-4ada-a94c-343e7085c1d7" name="Changes" comment="" />
+      <created>1667553296524</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1667553296524</updated>
+      <workItem from="1667553302784" duration="5290000" />
+      <workItem from="1668504154594" duration="62000" />
+      <workItem from="1668504314959" duration="62000" />
+      <workItem from="1668683578498" duration="211000" />
+      <workItem from="1669125064559" duration="875000" />
+      <workItem from="1670514503723" duration="621000" />
+      <workItem from="1670845494444" duration="14884000" />
+    </task>
     <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
     <option name="TAB_STATES">
