+++ src/App.js
@@ -10,6 +10,4 @@ function App() {
         </div>
     );
 }
-
-
 export default App;
