+++ src/components/UI/pagination/MyPagination.module.css
@@ -1,0 +1,3 @@
+.paginationAndRows {
+
+}
