+++ src/components/UI/search/MySearch.module.css
@@ -1,0 +1,3 @@
+.filters {
+
+}
